lkl tools: Make use of Kconfig for configuring the lkl host library.
Because tools/build currently has no Kconfig (or other reasonable config)
support, all %config targets and the include/config/auto.conf targets are
forwarded to the Kernel main makefile. Thus, the normal dependency
mechanisms kick in to properly configure the kernel or print an error
message if .config is missing.

Additionally, this patch removes the ability to compile the lkl host
libraries with a single command. "make" will no longer configure the
Kernel with the default config. You can, however, use "make defconfig" in
the tools/lkl directory to configure the kernel.

The Kconfig file in this patch adds support to select which tools and
library flavours are built.

Signed-off-by: Thomas Liebetraut <thomas@tommie-lie.de>