lkl: Complain more in virtio_process_one
Following the example of tools/lguest/lguest.c, a vring descriptor
chain that is too long to fit into our request is an error condition, so
we warn about it and panic when it happens.

This commit also adds some documentation to the very dense LKL-side virtio
code and splits up virtio_process_one into smaller functions.

Signed-off-by: Patrick Collins <pscollins@google.com>