lkl: Add operation to access host gettid
While LKL also holds on to an id for each thread it creates, we may want
to be able to inspect which thread we're running on at some point when
the LKL kernel is not in a good state --- for example, during netdev
setup, before the LKL kernel is initialized. This commit adds support
for accessing the host's notion of a tid, plus tests.

Signed-off-by: Patrick Collins <pscollins@google.com>