lkl: Allow output from special filesystems on exit
In order to debug networking performance issues, we'd like to be able to
see the state of e.g. /proc/net/netstat after a benchmark run. Since
there's no way to read from LKL's filesystem without lkl_sys_open,
getting this info makes it necessary to modify the benchmarking program
(e.g. netperf) to so that it performs the reads itself.

It's nontrivial to make this modification, so this commit provides a way
to access those filesystems via environment variables. First we can
specify a list of filesystem types we'd like:

    LKL_HIJACK_MOUNT=proc,sysfs

and they'll be mounted as /proc and /sysfs at startup. Then we specify
files we'd like to see:

    LKL_HIJACK_DUMP=/proc/net/snmp,/sysfs/class/net/eth0/mtu

and they're printed out when LKL exits.

Signed-off-by: Patrick Collins <pscollins@google.com>