lkl tools: Add tests for the VDE hijacking code
The added test for the hijacking using VDE network backend will only work
if the host has a working installation of vde_switch (contained in the vde2
tool suite). It creates a VDE network (consisting of a single switch) and
connects a host's tap device to it as first peer in the network.
The hijacked binaries will act as a second peer in the VDE network and test
LKL's interoperability with a known-working VDE implementation (VDE's own
tap device support).

Signed-off-by: Thomas Liebetraut <thomas@tommie-lie.de>