lkl tools: Implement virtio-net using VDE.
This commit adds an new lkl_netdev implementation using the vdeplug2 library
as a backend. This allows LKL to interface to the Virtual Distributed Ethernet
userspace network implementation and act as a client in a VDE network.

The VDE support is disabled by default and can be enabled by setting "vde=y" for the
makefile, e.g.
    $ make vde=y

Signed-off-by: Thomas Liebetraut <thomas@tommie-lie.de>