lkl: Change POSIX thread_create to match Win32
Although the documentation and the Win32 implementations say that the
return value of thread_create is a handle on the thread, the
implementation returns an error code instead. This change introduces a
new type for an LKL thread handle that can be used in both cases.

The current Win32 implementation is incorrect --- multiple places in the
codebase rely on thread_create returning 0 for success and nonzero for
failure, which have been fixed here. It is more convenient, however, to
change the POSIX implementation to match the Win32 implementation than
the other way around, because that gives us access to a thread handle
that we can use later on in helpful ways (e.g. adding a thread_join
host call).

Signed-off-by: Patrick Collins <pscollins@google.com>