lkl: Detach threads on POSIX to fix leak
On systems with pthreads, it's necessary to either join or detach all
threads in order to prevent a memory leak. This commit adds a host call
for pthread_detach only on POSIX, and it takes care of detaching the
thread that eventually becomes the default syscall thread, and all
kernel threads that are created. It is unnecessary to detach the threads
started via timer_create.

Signed-off-by: Patrick Collins <pscollins@google.com>