lkl tools: virtio net: use recursive queue locks
Since we can now run directed irqs it can happen that we run
virtio_process_queue twice from the same thread - first when the RX /
TX processing kicks in, and then as a result of an lkl_trigger_irq
which runs the interrupt directly. Hence, we need to convert the queue
locks to recursive locks.

Signed-off-by: Octavian Purdila <octavian.purdila@intel.com>