lkl: add support for direct syscalls
This adds support for running LKL system calls directly from the
calling host thread. We do this by "stealing" the LKL CPU when LKL is
idle. It is based on an idea and implementation by Yuan Liu
<liuyuan@google.com> but re-implemented to avoid modifying the core
schedule function.

This reduces the system call latency by an order of magnitude when LKL
is idle. It also removes the limitation of threads that can issue LKL
system calls (set to the number of IRQs available, which is 64 on
64bit systems).

The one caveat of the current implementation is that the idle loop
break is postponed until the idle thread is running again.

The implementation first makes sure we don't have any races with Linux
threads by attempting to take the CPU lock when we try to run a direct
system call.

After we have the lock, we need to setup the proper context for the
system call to run, by perform a context switch between the idle
thread and the target Linux task that we will use to run the system
call.

The reverse action needs to happen when we finished executing the
system call - we need to perform a context switch between the current
Linux task and the idle thread (or the thead that was picked by the
scheduler if other tasks were waked up in the process).

Because we now run system calls directly we don't need to have host
threads running associated with the Linux system call threads. As
such, part of the system call thread initialization, we terminate the
host threads (while we keep the Linux data structures in place).

This simplifies the system call implementation but complicates the way
context switching is done. We now have to alternate between multiple
types of context switching, based on the state we are in.

To make things cleaner we change the context switch implementation and
abstract the actions we need to perform on the next and prev
threads. Besides the default next and prev actions, we can have
specific next of prev actions for the following cases: host thread
"enter" - host thread has acquired the CPU and wants to run a system
call as a specific Linux thread, host thread "leave" - the reverse
operation, and host thread exit - terminate the host thread but keep
the Linux thread.

Signed-off-by: Octavian Purdila <octavian.purdila@intel.com>