lkl tools: virtio net fd: make it posix compatible
Rework the virtio net Linux fd implementation to make it POSIX
compatible. We do this by removing dependencies on epoll and eventfd.

Instead of using eventfd to shutdown the poll thread we use a control
pipe - closing the write side of the pipe will issue a POLLHUP and
wake-up the poll thread.

Note that this patch uses pipe2 instead of pipe in order to address a
libhijack limitation where pipe is hijacked regardless of calls coming
from the library or not. This makes virtio_net_fd Linux specific, but
we can easily switch back to pipe once the libhijack issue is resolved.

Like-wise, instead of using an epoll edge-triggered approach to
mitigate the high CPU usage on the TX side, we use a level-triggered
poll approach where the poll POLLIN / POLLOUT mask is set only when is
needed, i.e. when the read or write calls returns EAGAIN. The control
pipe is used to wake-up the poll thread and change the poll mask to
avoid race between the poll thread and LKL threads issuing reads or
writes.

Also collapse the rx and tx poll threads into a single poll thread. I
did not notice any performance degradation and reducing the number of
threads reduces the complexity.

Since the the poll thread can now be stopped reliably the patch also
removes lkl_netdevs_remove() and makes lkl_netdev_remove() available
to make the lkl_netdev_add/remove operations
symmetric. lkl_netdev_remove() can be called by the application before
lkl_sys_halt so we can also remove it from lkl_sys_halt.

Signed-off-by: Octavian Purdila <octavian.purdila@intel.com>