lkl tools: tests: add TAP13 output support
Using a standard test output makes it easier to show a unified output,
create reports for test runs and it allows for better CI integration.

TAP13 was chosen because it is a simple and human readable format while
still being pretty flexible when needed.

This patch adds test runner that support TAP13 output as well as a
simple test framework to make it easier to write and run tests from C
programs.

The main difference with old way of running tests is that the test
runner expects to receive the list of all tests before hand, in a test
array. This allows us catch hanged tests and generate meaninfull
reports.

Besides the standard output, the new test framework also adds a few
improvements: skip and bailout test results, execution measurement and
per test logging.

Signed-off-by: Octavian Purdila <tavi@cs.pub.ro>