Documentation/lkl.txt: fix Markdown formatting
Despite being a .txt file, this is symlinked as README.md at the root,
so must be supposed to be Markdown.  Throughout the file, triple
backticks (```) were used around Markdown code blocks, but those code
blocks were also indented.  Both of these are ways of making a
Markdown code block, but together they result in a top-level code
block containing code that is indented four spaces, which was probably
not the intention.  Additionally, the backticks were not themselves
indented, which had the effect of terminating the list they were
supposed to be contained in, resulting in markup like this:

  ...
  <li>
    <p>IPv6 gateway address</p>
    <p>key: "gateway6"<br>value type: string</p>
    <p>the gateway IPv6 address of the LKL network stack.</p>
  </li>
</ul>
<pre><code>
     "gateway6":"2001:db8:0:f101::1"
</code></pre>
<ul>
  <li>
    <p>Debug</p>
    ...

...which was clearly not the intention.

Signed-off-by: Alyssa Ross <hi@alyssa.is>